requirements-pi.txt
click==6.7
Flask==1.1.2
itsdangerous==0.24
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
picamera==1.13
python-dotenv
sense_hat
requests
